refactor to call function in different files
